+++ pkg/gateway/sys_windows.go
@@ -19,7 +19,7 @@ func setTCPFastOpen(fd uintptr) error {
 	return nil
 }
 
-func setCloExec(fd uintptr) error {
+func setCloExec(fd int) error {
 	return nil
 }
